+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.4 (xxxx-xx-xx)
+## 0.0.4 (2025-05-18)
 - **Processes**: added the ability to create multiple processes inside the OS
 - **Added PIT support**: PIT is a Programmable Interval Timer, like a clock simulation
 - **Added ATA driver**: this will be useful to save data permanently on a disk image